+++ pallets/offchain-task-scheduler/Cargo.toml
@@ -71,7 +71,6 @@ rev = "973dd744f5f7c6322799bbfa29366588e0267b0a"
 
 [features]
 default = ['std']
-runtime-benchmarks = ['frame-benchmarking']
 std = [
     'frame-support/std',
     'frame-system/std',
