+++ pallets/offchain-task-scheduler/src/tasks.rs
@@ -16,7 +16,7 @@ use sp_runtime::SaturatedConversion;
 
 #[inline]
 pub(crate) fn storage_key<Id: Encode>(id: &Id) -> Vec<u8> {
-	const TASK_GUARD: &[u8] = b"offchain-task-scheduler/task/guard/";
+	const TASK_GUARD: &[u8] = b"task-scheduler/task/guard/";
 	id.using_encoded(|encoded_id| TASK_GUARD.iter().chain(encoded_id).copied().collect())
 }
